fixed components and renamed files accordingly
